refactor: extract OperationsList component from OperationsScreen
Extracted FlatList rendering logic into a dedicated OperationsList component
to improve code organization and maintainability.

Changes:
- Created OperationsList component (269 lines)
  - Handles operations list rendering with date grouping
  - Manages lazy loading and scroll functionality
  - Renders date separators with spending sums
  - Renders individual operation items

- Updated OperationsScreen (1143 → 1011 lines, -132 lines, -11.5%)
  - Removed renderItem, renderFooter, handleEndReached callbacks
  - Kept groupedOperations memo in parent for scroll-to-date logic
  - Simplified component by delegating list rendering to OperationsList
  - Maintained all existing functionality (quick add, filters, modals)

Benefits:
- Improved separation of concerns
- Easier to test and maintain
- Reduced complexity of OperationsScreen
- Reusable list component for future features

Related to #1 in react-native-best-practices.md plan
